Renamed interfacebuilder.py to thread.py
