+++ src/SerialPlotter/panels/recorder.py
@@ -4,7 +4,7 @@ from typing import Dict
 
 from .. import mvc
 from ..filehandler import json_load, json_save, csv_save_append, csv_save_create
-from ..interfacebuilder import ThreadInterface
+from ..thread import ThreadInterface
 
 
 class View(mvc.View):
